third_party/blink/web_tests/register-protocol-handler/blob-urls.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<title>Test that blob: URLs fail registerProtocolHandler checks</title>
<body>
<iframe id='frame'></iframe>

<script>
const frameSrc =
`<html>
<head>
<script>
try {
  navigator.registerProtocolHandler('web+test',
                                    location.href + '?%s',
                                    'Test handler');
  window.parent.postMessage('rph succeeded', '*');
} catch (e) {
  window.parent.postMessage('rph failed: ' + e.message, '*');
}
<\/script>
<\/head>
<body>
Test page
<\/body>
<\/html>
`

window.addEventListener('message', (msg) => {
  assert_equals(msg.data,
                'rph failed: Failed to execute \'registerProtocolHandler\' on \'Navigator\': The scheme of the url provided must be the \'http\' or \'https\'.',
                'registerProtocolHandler should have failed');
  done();
});

const data = new Blob([frameSrc], {type: 'text/html'});
document.getElementById('frame').src = window.URL.createObjectURL(data);
</script>
</body>
